Update project after review
