+++ src/index.tsx
@@ -1,44 +1,11 @@
 import { createRoot } from 'react-dom/client';
-import { StrictMode, CSSProperties, useState } from 'react';
-import clsx from 'clsx';
-
-import { Article } from './components/article/Article';
-import { ArticleParamsForm } from './components/article-params-form/ArticleParamsForm';
-import { defaultArticleState, IFormData } from './constants/articleProps';
+import { StrictMode } from 'react';
 
 import './styles/index.scss';
-import styles from './styles/index.module.scss';
+import { App } from './app';
 
 const domNode = document.getElementById('root') as HTMLDivElement;
 const root = createRoot(domNode);
-
-const App = () => {
-	const inLocalStorage =
-		localStorage.getItem('form') !== null
-			? JSON.parse(localStorage.form)
-			: defaultArticleState;
-
-	const [state, setState] = useState<IFormData>(inLocalStorage);
-	const handleUpdate = (formData: IFormData) => {
-		setState(formData);
-	};
-	return (
-		<div
-			className={clsx(styles.main)}
-			style={
-				{
-					'--font-family': state.fontFamilyOption.value,
-					'--font-size': state.fontSizeOption.value,
-					'--font-color': state.fontColor.value,
-					'--container-width': state.contentWidth.value,
-					'--bg-color': state.backgroundColor.value,
-				} as CSSProperties
-			}>
-			<ArticleParamsForm isUpdate={handleUpdate} />
-			<Article />
-		</div>
-	);
-};
 
 root.render(
 	<StrictMode>
